Fix undefined sign-bit shifts in palette fades

+++ FistBlue/effects.c
@@ -94,9 +94,9 @@ static void adder(u16 **palbase, int *a1, short arg, char d7) {		//11ea
 	} else {
 		// 120a
 		if (arg < 0) {
-			*a1 |= (1 << d7);
+			*a1 |= (int)(1u << d7);
 		} else {
-			*a1 &= ~(1 << d7);
+			*a1 &= ~(int)(1u << d7);
 		}
 	}
 	*palbase += 16;
@@ -122,9 +122,9 @@ static void whiteadder(GPAL *palbase, int *fadebase, int count, short arg) {			/
 
 	d0 = *fadebase;
 	if (arg < 0) {
-		d0 |= ( 1 << 31);
+		d0 |= (int)(1u << 31);
 	} else {
-		d0 &= ~(1 << 31);
+		d0 &= ~(int)(1u << 31);
 	}
 	es.FadeCounter += d0;
 }
